Pi0.5: address PR review comments
- Refactor Pi05GemmaRMSNorm into shared components/__init__.py (was
  duplicated in paligemma.py). Both paligemma.py and action_expert.py
  now import from the shared location. (sivginirmak)
- Remove dead Pi05Tokenizer.encode_state() method and its unused
  discretize_state import. (merceod)
- Fix sincos_timestep_embedding to compute frequency progression in
  float64 (matching openpi reference) instead of t.dtype which is bf16
  under autocast. (merceod)
- Replace rigid load_pi05_config with auto-mapping: all Pi05Config
  fields are matched by name against HF config keys, with an explicit
  rename dict only for the few keys that differ (num_hidden_layers ->
  num_layers, etc.). (merceod)

All 17 unit tests pass.

+++ mminf/model/pi05/components/flow_matching.py
@@ -27,9 +27,12 @@ def sincos_timestep_embedding(
         t = t[None]
     half = dim // 2
     # Geometric progression of frequencies between min_period and max_period.
-    fraction = torch.linspace(0.0, 1.0, half, device=t.device, dtype=t.dtype)
+    # Use float64 for the frequency computation to match the openpi reference;
+    # bf16 has only ~3 digits of precision and rounds higher-frequency
+    # components, which compounds through time_mlp -> adaRMS -> 18 layers.
+    fraction = torch.linspace(0.0, 1.0, half, device=t.device, dtype=torch.float64)
     period = min_period * (max_period / min_period) ** fraction
-    omega = 2.0 * math.pi / period
+    omega = (2.0 * math.pi / period).to(t.dtype)
     angles = t[..., None] * omega
     return torch.cat([torch.sin(angles), torch.cos(angles)], dim=-1)
